requirements.txt
# bridge-analyst — runtime dependencies
# 設計原則：依賴越少越好，便於面試時在乾淨環境中重跑。

# --- Agent brain ---
anthropic>=0.40.0          # Claude SDK — agent 推理層

# --- 資料抓取（Day 2 scrapers） ---
httpx>=0.27.0              # 異步 HTTP（PTT / 新聞）
beautifulsoup4>=4.12.0     # HTML 解析
lxml>=5.0.0                # BS4 的快速後端
readability-lxml>=0.8.1    # 文章正文抽取（news scraper 的 fallback）
tenacity>=8.2.0            # 重試（PTT 18 禁頁 / 限速）

# --- 資料處理 ---
pandas>=2.0.0              # 輿情 dataframe 運算
python-dotenv>=1.0.0       # 讀 .env

# --- 視覺化與報告（Day 8+） ---
plotly>=5.20.0             # dashboard（跟 district-analyst 一致）
python-docx>=1.1.0         # Word 報告交付（主要格式）
reportlab>=4.0.0           # PDF 備用（Ch.1–Ch.4 預覽，須手動放 Noto Sans TC 字型）
matplotlib>=3.7.0          # charts
fonttools>=4.50.0          # 可選：若 build_pdf 需要 TTC 抽取

# --- CLI / DX ---
tqdm>=4.66.0               # scraper 進度條
tabulate>=0.9.0            # CLI 輸出表格

# --- 測試 ---
pytest>=8.0.0
